got a working ecosharder/expression maker Branch: feature/GEOPROCESSING-180-raster-calculator-expressions
